Correção da atividade de banco de dados com python

+++ main.py
@@ -24,7 +24,8 @@ class Aluno(Base):
     __tablename__ = "alunos"
 
     # Definindo atributos da tabela
-    ra = Column("ra", Integer, primary_key=True, autoincrement=True)
+    id = Column("id", Integer, primary_key=True, autoincrement=True)
+    ra = Column("ra", Integer)
     nome = Column("nome", String)
     idade = Column("idade", Integer)
     senha = Column("senha", String)
@@ -39,8 +40,6 @@ class Aluno(Base):
 # Criando tabela no banco de dados.
 Base.metadata.create_all(bind=db)
 
-os.system("cls || clear")
-
 # Salvar banco de dados
 for i in range(2):
      ra = input("Digite o seu ra: ")
@@ -53,16 +52,23 @@ for i in range(2):
      session.commit()
      print()
 
+os.system('cls || clear')
+
 print("\nListando alunos no banco de dados")
 lista_alunos = session.query(Aluno).all()
 
 for aluno in lista_alunos:
-    print(f"{aluno.ra} - {aluno.nome} - {aluno.idade} - {aluno.senha}")
+    print(f"{aluno.id} - {aluno.ra} - {aluno.nome} - {aluno.idade} - {aluno.senha}")
 
-aluno = session.query(Aluno).filter_by(nome="João").first()
-session.delete(aluno)
-session.commit()
-print("\nUsuário deletado com sucesso!")
+print("\nExcluindo alunos no banco de dados")
+ra_aluno = input("Informe o R.A do aluno: ")
+aluno = session.query(Aluno).filter_by(ra=ra_aluno).first()
+if aluno:
+    session.delete(aluno)
+    session.commit()
+    print("\nUsuário deletado com sucesso!")
+else:
+    print("\nUsuário não encontrado!")
 
 # usuario = Usuario("Ruan", "Ruan@gmail.com", "123")
 # aluno = Aluno(ra=222, nome="Ruan", idade=22, senha="123")
@@ -71,7 +77,39 @@ print("\nUsuário deletado com sucesso!")
 
 # Mostrando conteúdo do banco de dados.
 # SELECT * FROM Usuarios
+print("\nListando usuários do banco de dados")
 lista_alunos = session.query(Aluno).all()
 
 for aluno in lista_alunos:
-    print(f"{aluno.ra} - {aluno.nome}  - {aluno.idade} - {aluno.senha}")
+    print(f"{aluno.id} - {aluno.ra} - {aluno.nome}  - {aluno.idade} - {aluno.senha}")
+
+# Atualizar um aluno.
+print("\nAtualizando os dados de um usuário.")
+
+ra_aluno = input("Informe o R.A. do aluno: ")
+
+aluno = session.query(Aluno).filter_by(ra=ra_aluno).first()
+
+if aluno:
+    aluno.ra = input("Digite o seu R.A.: ")
+    aluno.nome = input("Digite o seu nome: ")
+    aluno.idade = input("Digite a sua idade: ")
+    aluno.senha = input("Digite a sua senha: ")
+    session.commit()
+else:
+    print("Usuário não encontrado!")
+
+# Pesquisando um usuário
+print("\nPesquisando um usuário pelo R.A. .")
+
+ra_aluno = input("Informe o R.A. do aluno: ")
+
+aluno = session.query(Aluno).filter_by(ra=ra_aluno).first()
+
+if aluno:
+    print(f"{aluno.id} - {aluno.ra} - {aluno.nome}  - {aluno.idade} - {aluno.senha}")
+else:
+    print("Usuário não encontrado!")
+
+#Fechando conexão.
+session.close()
